pyproject.toml
[project]
name = "audio-transcription-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "black>=25.11.0",
    "celery[redis]>=5.5.3",
    "fastapi[standard]>=0.121.2",
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.0",
    "openai>=2.8.0",
    "pytube>=15.0.0",
    "sqlalchemy>=2.0.44",
    "yt-dlp[default]>=2025.11.12",
]
